+++ ddtrace/contrib/_events/http_client.py
@@ -1,17 +1,15 @@
 from dataclasses import dataclass
 from enum import Enum
 from typing import Callable
-from typing import Mapping
-from typing import MutableMapping
 from typing import Optional
-from typing import Protocol
-from typing import Sequence
 from typing import Union
 
 from ddtrace._trace.events import TracingEvent
+from ddtrace.contrib._events.http import HttpBaseEvent
+from ddtrace.contrib._events.http import HttpRequestBaseEvent
+from ddtrace.contrib._events.http import _HttpClientResponse
 from ddtrace.ext import SpanKind
 from ddtrace.ext import SpanTypes
-from ddtrace.internal.core.events import Event
 from ddtrace.internal.core.events import event_field
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.schema import schematize_url_operation
@@ -21,43 +19,11 @@ from ddtrace.internal.schema.span_attribute_schema import SpanDirection
 log = get_logger(__name__)
 
 
-JsonType = Union[None, bool, int, float, str, Sequence["JsonType"], Mapping[str, "JsonType"]]
-
-
-class _HttpClientResponse(Protocol):
-    @property
-    def headers(self) -> MutableMapping[str, str]: ...
-    @property
-    def status_code(self) -> int: ...
-
-    def json(self) -> JsonType: ...
-
-
 class HttpClientEvents(Enum):
     HTTP_REQUEST = "http.client.request"
     HTTPX_REQUEST = "httpx.request"
     HTTP_SEND_REQUEST = "http.client.send_request"
     HTTPX_SEND_REQUEST = "httpx.send_request"
-
-
-@dataclass
-class HttpBaseEvent(Event):
-    request_url: str = event_field()
-    request_method: str = event_field()
-    request_headers: MutableMapping[str, str] = event_field()
-
-    response_headers: Mapping[str, str] = event_field(default_factory=dict)
-    response_status_code: Optional[int] = event_field(default=None)
-
-    def set_response(self, response: _HttpClientResponse) -> None:
-        self.response_status_code = response.status_code
-        self.response_headers = response.headers
-
-
-@dataclass
-class HttpRequestBaseEvent(HttpBaseEvent):
-    http_operation: str = event_field()
-    query: str = event_field()
 
 
 @dataclass
